# internal/swagger/swagger.yaml
basePath: /api
definitions:
  handler.Response:
    properties:
      code:
        description: Code 状态码，非0时表示自定义HTTP业务状态码
        type: integer
      data:
        description: Data 响应数据，具体内容因接口而异
      error_code:
        description: ErrorCode 业务错误码，可选
        type: string
      message_key:
        description: MessageKey 国际化消息 key，可选
        type: string
      message_params:
        additionalProperties: {}
        description: MessageParams 国际化模板参数，可选
        type: object
      msg:
        description: Msg 返回信息，通常是状态描述
        type: string
    type: object
  model.AccessTokenSettingDto:
    properties:
      audience:
        description: 访问令牌受众（public-client/cli/integration）
        type: string
      expiry:
        description: 访问令牌过期策略（8_hours/1_month/never）
        type: string
      name:
        description: 访问令牌名称
        type: string
      scopes:
        description: 访问令牌权限范围
        items:
          type: string
        type: array
    type: object
  model.AgentSettingDto:
    properties:
      api_key:
        description: LLM API Key
        type: string
      base_url:
        description: 自定义 API URL（可选）
        type: string
      enable:
        description: 是否启用 Agent 功能
        type: boolean
      model:
        description: LLM 模型名称
        type: string
      prompt:
        description: Agent 额外使用的提示词
        type: string
      protocol:
        description: LLM 接口协议（OpenAI 兼容/Anthropic/Gemini，OpenAI 兼容覆盖 DeepSeek、Qwen、Ollama
          等）
        type: string
    type: object
  model.BackupScheduleDto:
    properties:
      cron_expression:
        description: 备份计划的 Cron 表达式
        type: string
      enable:
        description: 是否启用备份计划
        type: boolean
    type: object
...
  model.EmbeddingSettingDto:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      dim:
        type: integer
      enable:
        type: boolean
      model:
        type: string
    type: object
...
  model.OAuth2SettingDto:
    properties:
      auth_redirect_allowed_return_urls:
        items:
          type: string
        type: array
      auth_url:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      cors_allowed_origins:
        items:
          type: string
        type: array
      enable:
        type: boolean
      is_oidc:
        description: 是否启用 OIDC
        type: boolean
      issuer:
        description: OIDC 颁发者
        type: string
      jwks_url:
        description: OIDC JWKS URL
        type: string
      provider:
        type: string
      redirect_uri:
        type: string
      scopes:
        items:
          type: string
        type: array
      token_url:
        type: string
      user_info_url:
        type: string
    type: object
...
  model.PasskeySettingDto:
    properties:
      webauthn_allowed_origins:
        items:
          type: string
        type: array
      webauthn_rp_id:
        type: string
    type: object
...
  model.S3SettingDto:
    properties:
      access_key:
        description: 访问密钥 ID
        type: string
      bucket_name:
        description: 存储桶名称
        type: string
      cdn_url:
        description: CDN 加速域名（可选，没有就走 Endpoint）
        type: string
      enable:
        description: 是否启用 S3 存储
        type: boolean
      endpoint:
        description: S3 端点
        type: string
      path_prefix:
        description: 存储路径前缀，例如 "uploads/"，方便隔离目录
        type: string
      provider:
        description: S3 服务提供商，例如 "aws", "aliyun", "minio", "other"
        type: string
      public_read:
        description: 上传时是否默认设置对象为 public-read
        type: boolean
      region:
        description: 区域
        type: string
      secret_key:
        description: 秘密访问密钥
        type: string
      use_ssl:
        description: 是否使用 SSL
        type: boolean
    type: object
...
  model.WebhookDto:
    properties:
      is_active:
        description: 启用/禁用状态
        type: boolean
      name:
        description: Webhook 名称
        type: string
      secret:
        description: 签名密钥，用于请求验证（HMAC等）
        type: string
      url:
        description: Webhook URL
        type: string
    type: object
host: localhost:6277
info:
  contact: {}
  description: 开源、自托管轻量级发布平台 Ech0 的 API 文档
  title: Ech0 API 文档
  version: "1.0"
paths:
  /access-tokens:
    get:
      consumes:
      - application/json
      description: 列出当前用户的所有访问令牌
      produces:
      - application/json
      responses:
        "200":
          description: 列出访问令牌失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 列出访问令牌
      tags:
      - 系统设置
    post:
      consumes:
      - application/json
      description: 为当前用户创建一个新的访问令牌
      parameters:
      - description: 新的访问令牌信息
        in: body
        name: accessToken
        required: true
        schema:
          $ref: '#/definitions/model.AccessTokenSettingDto'
      produces:
      - application/json
      responses:
        "200":
          description: 创建访问令牌失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 创建访问令牌
      tags:
      - 系统设置
  /access-tokens/{id}:
    delete:
      consumes:
      - application/json
      description: 根据 ID 删除指定的访问令牌
      parameters:
      - description: 要删除的访问令牌 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除访问令牌失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 删除访问令牌
      tags:
      - 系统设置
...
  /agent/settings:
    get:
      consumes:
      - application/json
      description: 获取系统的 Agent 相关设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取 Agent 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取 Agent 设置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 更新系统的 Agent 相关设置
      parameters:
      - description: 新的 Agent 设置
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/model.AgentSettingDto'
      produces:
      - application/json
      responses:
        "200":
          description: 更新 Agent 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新 Agent 设置
      tags:
      - 系统设置
...
  /backup/schedule:
    get:
      consumes:
      - application/json
      description: 获取系统的定期备份计划设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取备份计划失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取备份计划
      tags:
      - 系统设置
    post:
      consumes:
      - application/json
      description: 为系统设置定期备份计划
      parameters:
      - description: 备份计划设置
        in: body
        name: backupSchedule
        required: true
        schema:
          $ref: '#/definitions/model.BackupScheduleDto'
      produces:
      - application/json
      responses:
        "200":
          description: 设置备份计划失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新备份计划
      tags:
      - 系统设置
...
  /embedding/settings:
    get:
      consumes:
      - application/json
      description: 获取系统的 Embedding 向量相关设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取 Embedding 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取 Embedding 设置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 更新系统的 Embedding 向量相关设置
      parameters:
      - description: 新的 Embedding 设置
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/model.EmbeddingSettingDto'
      produces:
      - application/json
      responses:
        "200":
          description: 更新 Embedding 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新 Embedding 设置
      tags:
      - 系统设置
...
  /oauth2/settings:
    get:
      consumes:
      - application/json
      description: 获取系统的 OAuth2 相关设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取 OAuth2 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取 OAuth2 设置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 更新系统的 OAuth2 相关设置
      parameters:
      - description: 新的 OAuth 设置
        in: body
        name: oauthSettings
        required: true
        schema:
          $ref: '#/definitions/model.OAuth2SettingDto'
      produces:
      - application/json
      responses:
        "200":
          description: 更新 OAuth 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新 OAuth2 设置
      tags:
      - 系统设置
...
  /passkey/settings:
    get:
      consumes:
      - application/json
      description: 获取系统的 Passkey(WebAuthn) 相关设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取 Passkey 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取 Passkey 设置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 更新系统的 Passkey(WebAuthn) 相关设置
      parameters:
      - description: 新的 Passkey 设置
        in: body
        name: passkeySettings
        required: true
        schema:
          $ref: '#/definitions/model.PasskeySettingDto'
      produces:
      - application/json
      responses:
        "200":
          description: 更新 Passkey 设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新 Passkey 设置
      tags:
      - 系统设置
...
  /s3/settings:
    get:
      consumes:
      - application/json
      description: 获取系统的 S3 存储相关设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取 S3 存储设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取 S3 存储设置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 更新系统的 S3 存储相关设置
      parameters:
      - description: 新的 S3 存储设置
        in: body
        name: s3Settings
        required: true
        schema:
          $ref: '#/definitions/model.S3SettingDto'
      produces:
      - application/json
      responses:
        "200":
          description: 更新 S3 存储设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新 S3 存储设置
      tags:
      - 系统设置
  /settings:
    get:
      consumes:
      - application/json
      description: 获取系统的全局设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取设置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 更新系统的全局设置
      parameters:
      - description: 新的系统设置
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/model.SystemSettingDto'
      produces:
      - application/json
      responses:
        "200":
          description: 更新设置失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新设置
      tags:
      - 系统设置
...
  /webhook:
    get:
      consumes:
      - application/json
      description: 获取系统中配置的所有 Webhook 列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取 Webhook 列表失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取所有 Webhook
      tags:
      - 系统设置
    post:
      consumes:
      - application/json
      description: 创建一个新的 Webhook 配置
      parameters:
      - description: 新的 Webhook 配置
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/model.WebhookDto'
      produces:
      - application/json
      responses:
        "200":
          description: 创建 Webhook 失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 创建新的 Webhook
      tags:
      - 系统设置
  /webhook/{id}:
    delete:
      consumes:
      - application/json
      description: 根据 ID 删除指定的 Webhook 配置
      parameters:
      - description: 要删除的 Webhook ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除 Webhook 失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 删除 Webhook
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 根据 ID 更新指定的 Webhook 配置
      parameters:
      - description: 要更新的 Webhook ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 新的 Webhook 配置
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/model.WebhookDto'
      produces:
      - application/json
      responses:
        "200":
          description: 更新 Webhook 失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新 Webhook
      tags:
      - 系统设置
  /webhook/{id}/test:
    post:
      consumes:
      - application/json
      description: 根据 ID 触发一次 Webhook 测试请求
      parameters:
      - description: 要测试的 Webhook ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 测试 Webhook 失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 测试 Webhook
      tags:
      - 系统设置
...
swagger: "2.0"